overwriting mbr in c/c++ with zeros
how to overwrite Master Boot Record.   Master Boot Record is the first bytes on your hard disk, which has your boot data which is very important for the windows boot. it helps your computer to know how and where your OS is installled so that it can be loaded on your ram. if you overwrite Master Boot Record, This will render your windows unbootable. . do not use it for malicious purpose. and do not run this on actual hardware.